+++ changes.md
@@ -1,5 +1,9 @@
 #### Starting from 04/02/2023
 
+### 07/12/23
+- new file:   USACO/Guide/2-1-hammingCodes.cpp
+- modified:   git_add.py
+
 ### 07/11/23
 - new file:   CODEFORCES/1300/1844C.cpp
 - new file:   CODEFORCES/1400/1844D.cpp